requirements.txt
pandas==2.1.4
numpy==1.25.2
streamlit==1.28.1
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.7
schedule==1.2.0
openpyxl==3.1.2
apscheduler==3.10.4
openai==1.3.7
google-generativeai==0.5.2